Move test to main
