+++ internal/status/updater_test.go
@@ -55,6 +55,7 @@ var _ = Describe("Updater", func() {
 			Client:           client,
 			Logger:           zap.New(),
 			Clock:            fakeClock,
+			PodIP:            "1.2.3.4",
 		})
 	})
 
@@ -123,7 +124,8 @@ var _ = Describe("Updater", func() {
 			createExpectedGwWithGeneration = func(generation int64) *v1beta1.Gateway {
 				ipAddrType := v1beta1.IPAddressType
 				addr := v1beta1.GatewayAddress{
-					Type: &ipAddrType,
+					Type:  &ipAddrType,
+					Value: "1.2.3.4",
 				}
 
 				return &v1beta1.Gateway{
